split adif files out
