+++ ham_tools/adif/file.py
@@ -1,21 +1,14 @@
-"""
-ADIF utilities - parsing, merging, writing, etc.
-
-Reference: https://www.adif.org/adif
-Description of the file format: http://www.adif.org/312/ADIF_312.htm#ADI_File_Format
-"""
-
 import logging
-import re
 from collections import defaultdict
 from copy import copy
 from dataclasses import dataclass, field
-from datetime import date, datetime, time, timedelta
-from enum import Enum
+from datetime import datetime, timedelta
 from io import StringIO
 from pathlib import Path
 from typing import Optional, TextIO, Union
 
+from ham_tools.adif.record import AdifRecord, AdifSpecifier
+from ham_tools.adif.util import make_field, read_until
 from ham_tools.version import VERSION
 
 logger = logging.getLogger(__name__)
@@ -203,192 +196,3 @@ class AdifFile:
 
         # Sort all QSOs in this file by (date, time_on) ascending
         self.records = sorted(self.records, key=lambda r: (r.qso_date, r.time_on))
-
-
-@dataclass
-class AdifRecord:
-    # Raw fields, as a dict
-    fields: dict[str, str] = field(default_factory=dict)
-
-    _string_fields = {
-        "call",
-        "freq",
-        "gridsquare",
-        "my_gridsquare",
-        "tx_pwr",
-        "rst_rcvd",
-        "rst_sent",
-        "comment",
-    }
-
-    def __str__(self) -> str:
-        buf = ""
-        for k, v in self.fields.items():
-            buf += make_field(k, v) + " "
-        buf += "<eor>"
-        return buf
-
-    def __getitem__(self, key: str) -> str:
-        return self.fields[key]
-
-    def __setitem__(self, key: str, value: str) -> None:
-        self.fields[key] = value
-
-    def copy(self) -> "AdifRecord":
-        new = copy(self)
-        new.fields = self.fields.copy()
-        return new
-
-    def merge(self, other: "AdifRecord", force_overwrite: bool = False) -> None:
-        """
-        Merge another AdifRecord into this one. A field from the other record will
-        replace a field in this record if it is longer (likely better data), or
-        force_overwrite is set to True.
-        """
-        for k, v in other.fields.items():
-            set_field = k not in self.fields
-            set_field |= len(v) > len(self.fields.get(k, ""))
-            set_field |= force_overwrite
-
-            if set_field:
-                self.fields[k] = v
-
-    def _maybe_parse_date(self, field_name: str) -> Optional[date]:
-        f = self.fields.get(field_name)
-        if f:
-            return parse_date(f)
-        else:
-            return None
-
-    def _maybe_parse_time(self, field_name: str) -> Optional[time]:
-        f = self.fields.get(field_name)
-        if f:
-            return parse_time(f)
-        else:
-            return None
-
-    def _match_key(self) -> str:
-        """
-        Returns a string suitable as a dict key for telling if two QSOs likely match.
-        You need to also separately check that the times match up in addition.
-        """
-        parts = [
-            self.fields.get("callsign", "").upper(),
-            self.fields.get("band", "").lower(),
-            self.fields.get("mode", "").upper(),
-            self.fields.get("qso_date", ""),
-        ]
-        return "-".join(parts)
-
-    # Accessor methods which parse fields into Python-native types, e.g. dates and times
-    @property
-    def qso_date(self) -> Optional[date]:
-        return self._maybe_parse_date("qso_date")
-
-    @property
-    def time_on(self) -> Optional[time]:
-        return self._maybe_parse_time("time_on")
-
-    @property
-    def time_off(self) -> Optional[time]:
-        return self._maybe_parse_time("time_on")
-
-    @property
-    def datetime(self) -> Optional[datetime]:
-        """
-        Returns a datetime based on the qso_date and the time_on. Returns None if either
-        of those are None.
-        """
-        d = self.qso_date
-        t = self.time_on
-        if d is None or t is None:
-            return None
-
-        return datetime(
-            d.year, d.month, d.day, t.hour, t.minute, t.second, t.microsecond
-        )
-
-
-@dataclass
-class AdifSpecifier:
-    """
-    Specifier for a ADIF field, like <name:4>jawn
-
-    <field_name:length>
-    <field_name:length:type>
-    """
-
-    field_name: str
-    length: int = 0
-    type_: str = ""
-
-    @classmethod
-    def parse(cls, spec: str) -> "AdifSpecifier":
-        """
-        Parse a specifier from a string
-        """
-        spec = spec.lower().strip("<>")
-        parts = spec.split(":")
-
-        if len(parts) == 1:
-            # Something like <eoh> or <eor>
-            return AdifSpecifier(parts[0])
-        if len(parts) == 2:
-            # Just a name and a length
-            return AdifSpecifier(parts[0], int(parts[1]), "")
-        elif len(parts) == 3:
-            return AdifSpecifier(parts[0], int(parts[1]), parts[2])
-        else:
-            raise ValueError(f"Invalid specifier: {spec}")
-
-    @classmethod
-    def read_next(cls, f: TextIO) -> "AdifSpecifier":
-        """
-        Read the next specifier from the file. Will advance the position in the file.
-        """
-        read_until(f, "<")
-        buf = "<" + read_until(f, ">")
-        return cls.parse(buf)
-
-
-# Util functions
-def read_until(f: TextIO, char: str) -> str:
-    """
-    Returns the text from the current position in the file up to and including the given
-    char. If we hit the end of the file before finding our character, a ValueError is
-    raised.
-    """
-    buf = ""
-
-    cur = ""
-    while cur != char:
-        cur = f.read(1)
-        buf += cur
-        if cur == "":
-            raise ValueError(f"Hit end of file before finding '{char}'")
-
-    return buf
-
-
-def parse_date(date_str: str) -> date:
-    """
-    Parse an ADIF date - YYYYMMDD
-    """
-    return datetime.strptime(date_str, "%Y%m%d").date()
-
-
-def parse_time(time_str: str) -> time:
-    """
-    Parse an ADIF time - HHMM or HHMMSS
-    """
-    if len(time_str) == 4:
-        return datetime.strptime(time_str, "%H%M").time()
-    else:
-        return datetime.strptime(time_str, "%H%M%S").time()
-
-
-def make_field(name: str, value: str) -> str:
-    """
-    Return an ADIF field/value, like "<adif_ver:5>value"
-    """
-    return f"<{name}:{len(value)}>{value}"
